pyproject.toml
[tool.poetry]
name = "crypto-oracle-frontend"
version = "0.1.0"
description = ""
authors = ["lee <lee@logicon.io>"]

[tool.poetry.dependencies]
python = "^3.7"
"coinapi.rest.v1" = "^1.3"
Django = "3.1.6"
django-bootstrap-datepicker-plus = "3.0.5"
django-bootstrap4 = "2.3.1"
django-fontawesome-5 = "1.0.18"
django-model-utils = "4.1.1"
django-pandas = "0.6.4"
gunicorn = "20.0.4"
matplotlib = "3.3.4"
numpy = "1.19.5"
plotly = "4.14.3"
psycopg2-binary = "2.8.6"
python-dateutil = "2.8.1"
pytz = "2021.1"
requests = "2.23.0"
scipy = "1.6.0"
scikit-learn = "0.24.1"
boto3 = "^1.17.14"
ftx = "^1.0.2"
ciso8601 = "^2.1.3"
kucoin-python = "^1.0.5"
Werkzeug = "^1.0.1"

[tool.poetry.dev-dependencies]
coloredlogs = "^15.0"
django-debug-toolbar = "3.2"
django-extensions = "3.1.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
